+++ crates/core/keymap.rs
@@ -10,9 +10,12 @@ use std::{collections::HashMap, fmt::Debug};
 
 #[path = "keymap_codec.rs"]
 mod codec;
+#[path = "keymap_description.rs"]
+mod description;
 #[path = "keymap_defaults.rs"]
 mod defaults;
 
+pub use self::description::get_input_action_description;
 pub use self::defaults::{default_keymap, generate_keymap};
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd)]
@@ -323,140 +326,6 @@ pub enum InputAction {
     MouseMoveUp,
     #[serde(rename = "mouse_move_down")]
     MouseMoveDown,
-}
-
-pub fn get_input_action_description(input_action: InputAction) -> String {
-    match input_action {
-        // None
-        InputAction::None => "No action".to_string(),
-
-        // Up
-        InputAction::Up => "Move up".to_string(),
-        InputAction::WatchPaneUp => "Move up in watch pane".to_string(),
-        InputAction::HistoryPaneUp => "Move up in history pane".to_string(),
-
-        // Down
-        InputAction::Down => "Move down".to_string(),
-        InputAction::WatchPaneDown => "Move down in watch pane".to_string(),
-        InputAction::HistoryPaneDown => "Move down in history pane".to_string(),
-
-        // Shift Right
-        InputAction::ScrollRight => "Move Right".to_string(),
-        InputAction::ScrollHorizontalEnd => "Move Right end".to_string(),
-
-        // Shift Left
-        InputAction::ScrollLeft => "Move Left".to_string(),
-        InputAction::ScrollHorizontalHome => "Move Left home".to_string(),
-
-        // PageUp
-        InputAction::PageUp => "Move page up".to_string(),
-        InputAction::WatchPanePageUp => "Move page up in watch pane".to_string(),
-        InputAction::HistoryPanePageUp => "Move page up in history pane".to_string(),
-
-        // PageDown
-        InputAction::PageDown => "Move page down".to_string(),
-        InputAction::WatchPanePageDown => "Move page down in watch pane".to_string(),
-        InputAction::HistoryPanePageDown => "Move page down in history pane".to_string(),
-
-        // MoveTop
-        InputAction::MoveTop => "Move top".to_string(),
-        InputAction::WatchPaneMoveTop => "Move top in watch pane".to_string(),
-        InputAction::HistoryPaneMoveTop => "Move top in history pane".to_string(),
-
-        // MoveEnd
-        InputAction::MoveEnd => "Move end".to_string(),
-        InputAction::WatchPaneMoveEnd => "Move end in watch pane".to_string(),
-        InputAction::HistoryPaneMoveEnd => "Move end in history pane".to_string(),
-
-        // Focus
-        InputAction::ToggleFocus => "Toggle focus window".to_string(),
-        InputAction::FocusWatchPane => "Focus watch pane".to_string(),
-        InputAction::FocusHistoryPane => "Focus history pane".to_string(),
-
-        // Quit
-        InputAction::Quit => "Quit hwatch".to_string(),
-
-        // Reset
-        InputAction::Reset => "filter reset".to_string(),
-
-        // Delete
-        InputAction::Delete => "Delete selected history".to_string(),
-        InputAction::ClearExceptSelected => "Clear all history except selected history".to_string(),
-
-        // Cancel
-        InputAction::Cancel => "Cancel".to_string(),
-        InputAction::ForceCancel => "Cancel without displaying the exit dialog".to_string(),
-
-        // Help
-        InputAction::Help => "Show and hide help window".to_string(),
-
-        // Color
-        InputAction::ToggleColor => "Toggle enable/disable ANSI Color".to_string(),
-
-        // LineNumber
-        InputAction::ToggleLineNumber => "Toggle enable/disable Line Number".to_string(),
-
-        // Reverse
-        InputAction::ToggleReverse => "Toggle enable/disable text reverse".to_string(),
-
-        // Mouse Support
-        InputAction::ToggleMouseSupport => "Toggle enable/disable mouse support".to_string(),
-
-        // Toggle View Pane UI
-        InputAction::ToggleViewPaneUI => "Toggle view header/history pane".to_string(),
-        InputAction::ToggleViewHeaderPane => "Toggle view header pane".to_string(),
-        InputAction::ToggleViewHistoryPane => "Toggle view history pane".to_string(),
-
-        // Border
-        InputAction::ToggleBorder => "Toggle enable/disable border".to_string(),
-        InputAction::ToggleScrollBar => "Toggle enable/disable scroll bar".to_string(),
-        InputAction::ToggleBorderWithScrollBar => {
-            "Toggle enable/disable border and scroll bar".to_string()
-        }
-
-        // Diff Mode
-        InputAction::ToggleDiffMode => "Toggle diff mode".to_string(),
-        InputAction::SetDiffModePlane => "Set diff mode plane".to_string(),
-        InputAction::SetDiffModeWatch => "Set diff mode watch".to_string(),
-        InputAction::SetDiffModeLine => "Set diff mode line".to_string(),
-        InputAction::SetDiffModeWord => "Set diff mode word".to_string(),
-        InputAction::SetDiffOnly => "Set diff line only (line/word diff only)".to_string(),
-
-        // Output Mode
-        InputAction::ToggleOutputMode => "Toggle output mode".to_string(),
-        InputAction::SetOutputModeOutput => "Set output mode output".to_string(),
-        InputAction::SetOutputModeStdout => "Set output mode stdout".to_string(),
-        InputAction::SetOutputModeStderr => "Set output mode stderr".to_string(),
-
-        // Keyword search
-        InputAction::NextKeyword => "Focus next keyword".to_string(),
-        InputAction::PrevKeyword => "Focus previous keyword".to_string(),
-
-        // Toggle Wrap
-        InputAction::ToggleWrapMode => "Toggle wrap mode".to_string(),
-
-        // HistorySummary
-        InputAction::ToggleHistorySummary => "Toggle history summary".to_string(),
-
-        // Interval
-        InputAction::IntervalPlus => "Interval +0.5sec".to_string(),
-        InputAction::IntervalMinus => "Interval -0.5sec".to_string(),
-        InputAction::TogglePause => "Toggle Execution Pause".to_string(),
-
-        // Command/Filter
-        InputAction::ChangeFilterMode => "Change filter mode".to_string(),
-        InputAction::ChangeRegexFilterMode => "Change regex filter mode".to_string(),
-
-        // Mouse
-        InputAction::MouseScrollUp => "Mouse Scroll Up".to_string(),
-        InputAction::MouseScrollDown => "Mouse Scroll Down".to_string(),
-        InputAction::MouseButtonLeft => "Mouse Button Left".to_string(),
-        InputAction::MouseButtonRight => "Mouse Button Right".to_string(),
-        InputAction::MouseMoveLeft => "Mouse Move Left".to_string(),
-        InputAction::MouseMoveRight => "Mouse Move Right".to_string(),
-        InputAction::MouseMoveUp => "Mouse Move Up".to_string(),
-        InputAction::MouseMoveDown => "Mouse Move Down".to_string(),
-    }
 }
 
 #[cfg(test)]
